container: cleanup memory limit/guarantee
* memory_limit_total must work regardless of state

+++ src/util/unix.hpp
@@ -29,7 +29,7 @@ TError GetTaskChildrens(pid_t pid, std::vector<pid_t> &childrens);
 
 uint64_t GetCurrentTimeMs();
 bool WaitDeadline(uint64_t deadline, uint64_t sleep = 10);
-size_t GetTotalMemory();
+uint64_t GetTotalMemory();
 void SetProcessName(const std::string &name);
 void SetDieOnParentExit(int sig);
 std::string GetProcessName();
